guest-gcs-to-bq/requirements.txt
google-cloud-bigquery==2.32.0
google-cloud-storage